requirements.txt
aiofiles==0.4.0
beautifulsoup4==4.9.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
cymem==2.0.3
# en-core-web-md==2.2.5
fastapi==0.52.0
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.1
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.18.1
plac==1.1.3
preshed==3.0.2
pydantic==1.4
readability-lxml==0.7.1
regex==2020.4.4
requests==2.23.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
starlette==0.13.2
thinc==7.4.0
tqdm==4.46.0
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
wasabi==0.6.0
websockets==8.1
zipp==3.1.0
